requirements.txt
sentence-transformers
faiss-cpu
pdfplumber
pandas
numpy
